Qt: Don't change the UAT tree visibility when resizing
It's not necessary to change the visibility of the UAT tree view
when resizing the columns to the contents, and setting the visiblity
to false has the side effect of making any editor that is still
open not editable.

This mostly affects the PathSelectionEdit used by directory and file
names, which pass updates to the UAT after each character entered,
instead of only after pressing Enter. (The UAT code does validation,
but some widgets do their own validation too.)

This prevents having to click in the editor after each character
when entering directories or filenames.