Qt: ATapDataModel rows don't have children
Our ATapDataModel is table based, and the rows don't have children.
It's displayed in a TrafficTree, which inherits from QTreeView, however,
which allows the possibility that rows have children.

Make sure that rowCount returns 0 if handed a valid parent QModelIndex.

See:
https://doc.qt.io/qt-6/qabstractitemmodel.html#rowCount

We don't call rowCount with a valid index directly in our code, but some
of the Qt code can do so, for example QTreeView::expandRecursively() if
an Asterisk is pressed.

Fix #18975