MPEG PES: Update a comment
Explain where all the values for the stream_id are described in
ISO/IEC 13818-1 / ITU-T H.222.0 and ISO/IEC 13818-2 / ITU-T H.262.

Some of the values are not in fact stream_id values and not at the
PES level, but instead found in the MPEG video bitstream at a
different level of the hierarchy. The video bitstream and the PES
packets use disjoint set of values for the start codes so that
resynchronization in the face of errors doesn't suffer from ambiguity.