wmem: Allow integer lookups with a null tree
wmem_tree currently returns NULL for lookups of strings with
a NULL tree, but tries to dereference the tree and crashes when
looking up integers with a NULL tree. Return NULL (or FALSE, for
contains32()) for a tree that doesn't exist, making it easier to
lazily create a tree only when inserting an entry.