Decouple the media_type dissector table from HTTP.
Put the header containing the structure that should be passed to
dissectors in the media_type dissector table, and the code that creates
the media_type dissector table, into epan/dissectors/packet-media-type.h
and epan/dissectors/packet-media-type.h, respectively.

Change the names of types, enum values, structure fields, and variables
to speak of "media containers" and "content" rather than to speak of
HTTP or messages; HTTP is not the only protocol or data format that can
contain media type headers and data whose type is specified by media
type headers.