DRDA: Store TYPDEFNAM, CCSIDs as conversation data, implement SQLCARD
Store the TYPDEFNAM and TYPDEFOVR CCSIDs values (single and mixed byte)
as transmitted in ACCRDB and ACCRDBRM commands as conversation data
for the side of the connection. Track server versus client direction.

Handle when values are overridden for a command (including later command
data objects in the command chain with matching correlation number).

Store the negotiated SQL Application Manager level from the EXCSATRM
comand.

Have preferences for default values for when a capture does not include
the ACCRDB[RM] and EXCSATRM commands.

Implement generic handling of the FD:OCA data types necessary for
SQLSTT and SQLCARD, using the values from TYPDEFNAM and TYPDEFOVR.

Implement SQLCARD, handling the two different possible structures
depending on whether the SQLAM Level is less than 7 or greater than
equal to 7.

Implement the RLSCONV command.

Fix #18972