Qt: Sort display filter combobox by MRU
Sort the model by Most Recently Used, instead of Most Recently
Added. This improves the usability of the combobox and prevents
the most used filters from being pushed out of the list by newer
entries.

Add a monotonic increasing timestamp to each row, set it
when each entry is created or updated, and use it to sort the model.

Fixes #18997.