sFlow: Fix parsing of TCP flags and IPv4 ToS for sampled IP
TCP flags and IPv4 terms of service (ToS) fields
of sampled IP structures of sFlow are 32-bit big-endian
holding 8 meaningful bits each (in the last octet of 4).
The dissector treated them as little-endian
and added to parse as 8-bit fields (in the first octet of 4).
As a result, no TCP flag or ToS bit was ever reported set.

Parse the fields above according to the specification [1].
Wireshark mistreating them has been confirmed in the mailing list [2].

[1]: https://sflow.org/SFLOW-STRUCTS5.txt
[2]: https://groups.google.com/g/sflow/c/kqF7LRCVOBY