pgsql: Move state transitions to a tree
PostgreSQL has a number of packets whose interpretation depends
on state, especially the most recent packet sent from the other side.
Use a wmem_tree in the conversation data to store the state transitions
instead of just storing the last state value, because the latter only
works in sequential processing.

Add some extra items to the tree to indicate that we did get PGSQL
when it's a SSLResponse or a GSSENCResponse, and items for the request
codes so that bytes in the payload are fully covered.

If we haven't seen any PostgreSQL yet and it doesn't look like a
Postgres packet (the first byte is not ASCII nor the null byte),
reject the packet. It could be PGSQL over TLS, where the capture
started after the STARTTLS.

Only look for SSLResponse on the other side as the request.

Fix #19022