prefs: Describe converted pref names properly
The Column prefs were moved to a submodule, so the name needs to
be obtained from the parent module instead of the module found,
unlike other modules.

The Column preference names themselves were also changed in addition
to the module. Wait until checking both the module name and the
preference name before printing a warning to the user that a preference
has been converted to a new name, so that the correct name is displayed.

This also means that we print the warning message now for the preferences
that just had their names changed, not the modules.

Fix #19043