Qt: Make the debounce timer an advanced preference
Remove the debounce timer from the Appearance settings pane. The default
setting should be good enough for most people and the setting is still
available in the "Advanced" section if not.

Add comments to main_window_preferences_frame.ui noting that the
Language setting should be the last item, as in "if Wireshark is in a
language you can't read, go to this physical location in the UI."