Qt: Avoid unnecessary retaps when creating RTP and VOIP dialogs
Similar to 5bc94bc02a9f006a5eec504506a147f48d3101eb. Don't use
the automatic connection set up by Qt uic but manually connect
the display filter checkbox signal to its slot. This allows us
to initialize the checkbox and filter status without calling the
slot and retapping the packets in the middle of constructor.

There's always an tap triggered at the end of the constructor,
so we don't want to do a second one before the dialog appears if
the main display filter is set.