RTP: Have dynamic payload type show up as new Decode As item
The Decode As dialog, when adding a new item, auto-populates
with the data from the last Decode As item to appear in the frame.
This is because usually the presence of a Decode As item indicates
that previous protocols were dissected correctly.

The RTP protocol obscures the dynamic payload types because they
weren't presented in the prompt for the old GTK+ Decode As widget.
The prompts aren't used at all in Qt, but obscuring the dynamic
payload type means that the Decode As model and delegate don't
present dynamic payload types as something that can be changed in
the current packet.

Change that, so if you select Decode As on a packet that has, e.g.
RTP over UDP, it will prefer changing the RTP payload type dissector
to changing the dissector for the UDP port.