Qt: Simplify Decode As handling
We want to default to the last Decode As table with an entry in
the packet. Since the list of layers is a doubly-linked list,
start with the tail and go backwards for simpler and faster logic.

Move the logic for how to create a DecodeAsItem from a decode_as_t
into the constructors.