Qt: Decode As: Update the default dissector string
When changing a Decode As row item's table or selector, update
the default dissector string to refer to the default for the
new value.

Make access to the table names and selectors in a DecodeAsItem
through getters and setters, keeping the elements private, so
that we can ensure this happens whenever the table or selector
is changed.