Qt: Only have a getter for the default dissector
The DecodeAsItem determines the default dissector programmatically
when the table or value is changed. Other classes don't need to update it.

There's a value for the default dissector for a table and value written to
the decode_as_entries configuration file, but that has never affected
anything, because the actual default dissector is used. It is only
useful for information when inspecting or viewing the file.