SNMP: Fix multiple PDU over TCP handling
SNMP over TCP isn't common, but there is an implementation in
the dissector. It doesn't use tcp_dissect_pdus because of the ASN.1
parsing.

Fix a typo when there are multiple PDUs in a single frame - the
new offset needs to be passed so that later PDUs are dissected.
Also the return value when there isn't enough data to determine
the necessary length needs to be 0, not -1, due to how the SNMP
dissector handles it before handing back to TCP.

Fix some issues seen in fuzzed files.