MySQL: Make random access dissection work
Various data that is currently tracked in conversation data needs to
be moved to the per PDU data. The conversation data is only for the
last encountered value.

The remaining field packet count needs to be tracked in per frame data.

The most recent field metadata needs to be tracked in per frame data.

The number of parameters and number of fields received during the
last prepared statement can be retrieved from the statement data
of the last prepared statement id, which is already in per PDU data.

The statement data for a given statement id (which is not reused on
the same connection) should only be created on the first pass instead
of being recreated (in file scope) each time a PREPARE PDU is dissected.