GitLab CI: Check the ASN.1 dissectors consistency
Regenerate the ASN.1 dissectors in the Code Checks job, and fail
if any tracked files have changes after doing so.

This should make it impossible to commit changed to ASN.1 dissectors
without corresponding changes to the templates and conformance files,
and vice versa. (Perhaps the generated files shouldn't even be in git.)