MySQL: Handle resultset rows in text format beginning with 0
A binary resultset row always begins with a 0 header byte, while
a text resultset row in text format begins with a length encoded
string of the first column directly.

For most column values this is enough to distinguish the two, but
if the first value is an empty string (not a NULL value, encoded as
0xFB), then the text format also begins with a 0. In that case, we
have to track whether the request used the text protocol (COM_QUERY, the
deprecated COM_PROCESS_INFO) or the binary protocol (COM_STMT_EXECUTE,
COM_STMT_FETCH).

See https://mariadb.com/kb/en/resultset-row/
https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_binary_resultset.html#sect_protocol_binary_resultset_row
https://dev.mysql.com/doc/dev/mysql-server/latest/page_protocol_com_query_response_text_resultset_row.html