NSIS+MinGW: Add uninstaller
We remove the workaround for signing an installer that requires
creating an installer that just exists to write an uninstaller
and use WriteUninstaller instead to generate the installer.

This is done at install time. This method allows the uninstaller
to be created portably, namely using Linux.

To add back the signing aspect, including for Windows, we need
to use !uninstfinalize. This changes the compiler to generate
the installer at compile time and is also a portable method of
signing code. The code signing script will have to be platform
specific. This is called by the NSIS compiler using system()
and must be a batch file on Windows. The code assumes a script
named 'sign-wireshark.bat' exists on the path to sign the
executables.

An example of a sign.bat script would be:
    signtool sign /f c:\path\to\MySPC.pfx /t http://[insert timestamp URL] /fd sha256 %1

This is also a nice simplification of the build procedure for NSIS.

The !uninstfinalize stanza requires NSIS 3.08.