tshark: Document -T fields escaping, allow it to be turned off
Allow the escaping of whitespace characters and backspace with
the -T fields options to be disabled. There may be some use
cases (particular for redirected output instead of viewing at
a terminal) for not escaping, particularly since escaping makes
it difficult to distinguish a literal "\n" from an escaped newline.

Document this option, which also documents the escaping behavior.

Also add vertical tab to the list of escaped characters, for the
same reason as the others.

Fix #15796