wiretap: Generate IDBs from packets when necessary
Add a routine to generate a dummy IDB from a packet record.

When pcapng is writing enhanced packet blocks and the source
doesn't provide an interface id, search through the list of
intereface ids for a match. If there isn't one, generate a new
one and use it.

This allows pcapng to write per-packet encapsulation when the
source doesn't provide IDBs.